+++ truffle/src/com.oracle.truffle.object/src/com/oracle/truffle/object/LayoutImpl.java
@@ -40,6 +40,7 @@
  */
 package com.oracle.truffle.object;
 
+import java.lang.invoke.MethodHandles;
 import java.lang.invoke.VarHandle;
 import java.util.Map;
 import java.util.Objects;
@@ -133,10 +134,22 @@ public abstract class LayoutImpl extends com.oracle.truffle.api.object.Layout {
      * Preinitializes DynamicObject layouts for native image generation.
      *
      * NOTE: this method is called reflectively by downstream projects.
+     *
+     * @deprecated method overload for JDK 21 backwards compatibility.
      */
+    @Deprecated(since = "24.2")
     static void initializeDynamicObjectLayout(Class<?> dynamicObjectClass) {
+        initializeDynamicObjectLayout(dynamicObjectClass, null);
+    }
+
+    /**
+     * Preinitializes DynamicObject layouts for native image generation.
+     *
+     * NOTE: this method is called reflectively by downstream projects.
+     */
+    static void initializeDynamicObjectLayout(Class<?> dynamicObjectClass, MethodHandles.Lookup lookup) {
         assert TruffleOptions.AOT : "Only supported during image generation";
-        ((CoreLayoutFactory) getFactory()).registerLayoutClass(dynamicObjectClass.asSubclass(DynamicObject.class));
+        ((CoreLayoutFactory) getFactory()).registerLayoutClass(dynamicObjectClass.asSubclass(DynamicObject.class), lookup);
     }
 
     @SuppressWarnings("static-method")